moving over from old issue-277

+++ ui/game/functions.lua
@@ -98,27 +98,16 @@ function G.FUNCS.toggle_players_jokers()
 		local your_jokers_save = copy_table(G.jokers:save())
 		MP.end_game_jokers:load(your_jokers_save)
 
-		if MP.end_game_jokers.cards then
-			for _, card in pairs(MP.end_game_jokers.cards) do
-				card.mp_end_game_display = nil
-			end
+		for _, card in pairs(MP.end_game_jokers.cards or {}) do
+			card.mp_end_game_display = true
 		end
 
 		MP.end_game_jokers_text = localize("k_your_jokers")
 	else
 		if MP.end_game_jokers_received then
 			G.FUNCS.load_end_game_jokers()
-
-			if MP.end_game_jokers.cards then
-				for _, card in pairs(MP.end_game_jokers.cards) do
-					card.mp_end_game_display = true
-					card.added_to_deck = false
-				end
-			end
 		else
-			if MP.end_game_jokers.cards then
-				remove_all(MP.end_game_jokers.cards)
-			end
+			if MP.end_game_jokers.cards then remove_all(MP.end_game_jokers.cards) end
 			MP.end_game_jokers.cards = {}
 		end
